+++ packages/provider-elasticsearch/CHANGELOG.md
@@ -1,5 +1,6 @@
 # 1.12.0
 * [pivot] Add sort api
+* [pivot] Add filter api
 * [fieldValuesGroupStats] `.value` not needed for sorting
 * [fieldValuesGroupStats]: `sort.order` renamed to `sort.direction`
 * [numberRangesGroupStats] Fix drilldown logic to be lt instead of lte